chore: extract out balance calculation logic
